Context added for checking for mobile devices
