+++ src/About/About.tsx
@@ -21,7 +21,7 @@ import { Typography, Collapse } from "@material-ui/core";
 import GoldDivider from "../Utils/GoldDivider";
 import GreenLink from "../Utils/GreenLink";
 import ScrollToTopButton from "../Utils/ScrollToTopButton";
-import { useCurrentWidth } from "react-socks";
+import { MobileContext } from "../Context/MobileContext";
 
 const officers = [
     {
@@ -95,74 +95,76 @@ function About() {
 
     const [showAboutWebsite, setShowAboutWebsite] = React.useState<boolean>(false);
 
-    const mobile = useCurrentWidth() < 1000;
-
     DocumentTitle({ title: "About" });
 
     return (
-        <PageFade>
-            <div>
-                <Slideshow />
+        <MobileContext.Consumer>
+            {mobile => (
+                <PageFade>
+                    <div>
+                        <Slideshow />
 
-                <AboutSection
-                    title={"Officers"}
-                    subtitle={"Our officers help run the club so that members only need to worry about working on the robot"}
-                    content={officers}
-                />
+                        <AboutSection
+                            title={"Officers"}
+                            subtitle={"Our officers help run the club so that members only need to worry about working on the robot"}
+                            content={officers}
+                        />
 
-                <AboutSection
-                    title={"Web Team"}
-                    subtitle={"Our web team works diligently to provide a stunning website to showcase the club"}
-                    content={webTeam}
-                />
+                        <AboutSection
+                            title={"Web Team"}
+                            subtitle={"Our web team works diligently to provide a stunning website to showcase the club"}
+                            content={webTeam}
+                        />
 
-                <ContentBackground elevation={24} className="about-website-container">
-                    <Typography align="center" variant="h2" >
-                        About the Website
-                    </Typography>
-                    <GoldDivider />
-                    <ScrollToTopButton
-                        onClick={() => setShowAboutWebsite(!showAboutWebsite)}
-                        tooltipText={showAboutWebsite ? "Show Less" : "Show More"}
-                        style={{
-                            display: "flex",
-                            transition: "transform 0.5s",
-                            transform: !showAboutWebsite ? "rotate(180deg)" : "",
-                            padding: mobile ? "6px" : "",
-                            margin: "auto",
-                        }}
-                    />
-                    <Collapse in={showAboutWebsite} unmountOnExit>
-                        <Typography style={{ lineHeight: mobile ? "2.5rem" : "3.5rem" }} align="center" variant="h4">
-                            This website was built using
-                            {aboutWebsiteLinks.map((website) => {
-                                return (
-                                    <React.Fragment key={website.link}>
-                                        <br />
-                                        <GreenLink placement={"right"} link={website.link} text={website.name} />
-                                    </React.Fragment>
-                                );
-                            })}
-                            <br /><br />
-                            This website started as a semester project done by the students listed above in the Web Team
-                            section. We began with a flat white background and some randomly generated Latin text. From there,
-                            each member worked throughout the semester to add content and learn about the technologies used
-                            to build this site. The website at the end of the semester can be seen{" "}
-                            <GreenLink link="http://car-robotics.s3-website.us-east-2.amazonaws.com/" text="here" />.
-                            <br /><br />
-                            Since the end of that semester, I, the Matt in the Web Team section, continued working to refine each
-                            page for styling consistency, content, and optimizing user experience.<br />
-                            <br />
-                            The code for this website is open source and available{" "}
-                            <GreenLink link="https://github.com/car-robotics/CAR-Website-REACT" text="here" />
-                            {" "}for anyone to use for their own website.
-                            For those looking to learn React, this could be a good starting point as it gives you something to
-                            reference when styling/coding your own components.
-                        </Typography>
-                    </Collapse>
-                </ContentBackground>
-            </div>
-        </PageFade>
+                        <ContentBackground elevation={24} className="about-website-container">
+                            <Typography align="center" variant="h2" >
+                                About the Website
+                                </Typography>
+                            <GoldDivider />
+                            <ScrollToTopButton
+                                onClick={() => setShowAboutWebsite(!showAboutWebsite)}
+                                tooltipText={showAboutWebsite ? "Show Less" : "Show More"}
+                                style={{
+                                    display: "flex",
+                                    transition: "transform 0.5s",
+                                    transform: !showAboutWebsite ? "rotate(180deg)" : "",
+                                    padding: mobile ? "6px" : "",
+                                    margin: "auto",
+                                }}
+                            />
+                            <Collapse in={showAboutWebsite} unmountOnExit>
+                                <Typography style={{ lineHeight: mobile ? "2.5rem" : "3.5rem" }} align="center" variant="h4">
+                                    This website was built using
+                                        {aboutWebsiteLinks.map((website) => {
+                                        return (
+                                            <React.Fragment key={website.link}>
+                                                <br />
+                                                <GreenLink placement={"right"} link={website.link} text={website.name} />
+                                            </React.Fragment>
+                                        );
+                                    })}
+                                    <br /><br />
+                                    This website started as a semester project done by the students listed above in the Web Team
+                                    section. We began with a flat white background and some randomly generated Latin text. From there,
+                                    each member worked throughout the semester to add content and learn about the technologies used
+                                        to build this site. The website at the end of the semester can be seen{" "}
+                                    <GreenLink link="http://car-robotics.s3-website.us-east-2.amazonaws.com/" text="here" />.
+                                        <br /><br />
+                                    Since the end of that semester, I, the Matt in the Web Team section, continued working to refine each
+                                        page for styling consistency, content, and optimizing user experience.<br />
+                                    <br />
+                                    The code for this website is open source and available{" "}
+                                    <GreenLink link="https://github.com/car-robotics/CAR-Website-REACT" text="here" />
+                                    {" "}for anyone to use for their own website.
+                                    For those looking to learn React, this could be a good starting point as it gives you something to
+                                    reference when styling/coding your own components.
+                                    </Typography>
+                            </Collapse>
+                        </ContentBackground>
+                    </div>
+                </PageFade>
+            )}
+        </MobileContext.Consumer >
     );
 }
