+++ src/About/OfficerCard.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Card, CardActionArea, CardMedia, CardContent, Typography, makeStyles, Theme, createStyles, Divider } from "@material-ui/core";
 import EasyToSeeTooltip from "../Utils/EasyToSeeTooltip";
 import { COLORS } from "../Utils/COLORS";
-import { useCurrentWidth } from "react-socks";
+import { MobileContext } from "../Context/MobileContext";
 
 export enum Officer {
     President = "President",
@@ -40,33 +40,36 @@ const useStyles = makeStyles((theme: Theme) =>
 );
 
 export default function OfficerCard(props: OfficerCardProps) {
-    const mobile = useCurrentWidth() < 1000;
     const classes = useStyles();
     return (
-        <EasyToSeeTooltip title={props.linkedIn}>
-            <Card
-                className={classes.officerCard}
-                style={{
-                    flexBasis: mobile ? "" : "20%",
-                    margin: mobile ? "1rem auto" : "1rem",
-                }}
-            >
-                <CardActionArea disableRipple focusRipple={false} href={props.linkedIn} target="_blank">
-                    <CardMedia
-                        image={props.image}
-                        className={classes.officerPicture}
-                    />
-                    <CardContent>
-                        <Typography align="center" variant="h4" style={{ color: COLORS.darkColor }}>
-                            {props.officer}
-                        </Typography>
-                        <Divider />
-                        <Typography align="center" variant="h5" style={{ color: COLORS.darkColor }}>
-                            {props.name}
-                        </Typography>
-                    </CardContent>
-                </CardActionArea>
-            </Card>
-        </EasyToSeeTooltip >
+        <MobileContext.Consumer>
+            {mobile => (
+                <EasyToSeeTooltip title={props.linkedIn}>
+                    <Card
+                        className={classes.officerCard}
+                        style={{
+                            flexBasis: mobile ? "" : "20%",
+                            margin: mobile ? "1rem auto" : "1rem",
+                        }}
+                    >
+                        <CardActionArea disableRipple focusRipple={false} href={props.linkedIn} target="_blank">
+                            <CardMedia
+                                image={props.image}
+                                className={classes.officerPicture}
+                            />
+                            <CardContent>
+                                <Typography align="center" variant="h4" style={{ color: COLORS.darkColor }}>
+                                    {props.officer}
+                                </Typography>
+                                <Divider />
+                                <Typography align="center" variant="h5" style={{ color: COLORS.darkColor }}>
+                                    {props.name}
+                                </Typography>
+                            </CardContent>
+                        </CardActionArea>
+                    </Card>
+                </EasyToSeeTooltip >
+            )}
+        </MobileContext.Consumer>
     );
 }
