+++ src/Main/Main.scss
@@ -29,3 +29,7 @@
     text-align: center;
     overflow: hidden;
 }
+
+.footer-section {
+    @include checkMobile(width, "", 20%);
+}
