+++ src/Outreach/EventSection.tsx
@@ -5,7 +5,7 @@ import "./Outreach.scss";
 import { ContentBackground } from "../Utils/ContentBackground";
 import GoldDivider from "../Utils/GoldDivider";
 import { COLORS } from "../Utils/COLORS";
-import { useCurrentWidth } from "react-socks";
+import { MobileContext } from "../Context/MobileContext";
 
 interface EventSectionProps {
     eventName: string;
@@ -16,45 +16,47 @@ interface EventSectionProps {
 
 export default function EventSection(props: EventSectionProps) {
 
-    const mobile = useCurrentWidth() < 1000;
-
     return (
-        <div style={{ margin: mobile ? "3rem 0" : "4rem 0" }}>
-            <ContentBackground elevation={24} className="containers" style={{ padding: mobile ? "1rem 0rem" : "" }}>
-                <Typography align="center" style={{ padding: "0 1rem" }} variant="h2">
-                    {props.eventName}
-                </Typography>
-                <GoldDivider />
-                <div className="img">
-                    {props.images.map((img) => {
-                        return (
-                            <Image
-                                key={img}
-                                onClick={() => props.onClick(img)}
-                                style={{
-                                    textAlign: "center",
-                                    backgroundColor: "transparent",
-                                    paddingTop: "0",
-                                }}
-                                imageStyle={{
-                                    position: "inherit",
-                                    width: mobile ? "20rem" : "40rem",
-                                    height: mobile ? "15rem" : "30rem",
-                                    margin: "0.75rem",
-                                    boxShadow: `0px 0px 10px ${COLORS.schoolGold}`,
-                                    borderRadius: "0.5rem",
-                                }}
-                                aspectRatio={4 / 3}
-                                src={img}
-                            />
-                        );
-                    })}
+        <MobileContext.Consumer>
+            {mobile => (
+                <div style={{ margin: mobile ? "3rem 0" : "4rem 0" }}>
+                    <ContentBackground elevation={24} className="containers" style={{ padding: mobile ? "1rem 0rem" : "" }}>
+                        <Typography align="center" style={{ padding: "0 1rem" }} variant="h2">
+                            {props.eventName}
+                        </Typography>
+                        <GoldDivider />
+                        <div className="img">
+                            {props.images.map((img) => {
+                                return (
+                                    <Image
+                                        key={img}
+                                        onClick={() => props.onClick(img)}
+                                        style={{
+                                            textAlign: "center",
+                                            backgroundColor: "transparent",
+                                            paddingTop: "0",
+                                        }}
+                                        imageStyle={{
+                                            position: "inherit",
+                                            width: mobile ? "20rem" : "40rem",
+                                            height: mobile ? "15rem" : "30rem",
+                                            margin: "0.75rem",
+                                            boxShadow: `0px 0px 10px ${COLORS.schoolGold}`,
+                                            borderRadius: "0.5rem",
+                                        }}
+                                        aspectRatio={4 / 3}
+                                        src={img}
+                                    />
+                                );
+                            })}
+                        </div>
+                        <GoldDivider />
+                        <Typography className="supporting-text" align="center" variant="h5" >
+                            {props.supportingText}
+                        </Typography>
+                    </ContentBackground>
                 </div>
-                <GoldDivider />
-                <Typography className="supporting-text" align="center" variant="h5" >
-                    {props.supportingText}
-                </Typography>
-            </ContentBackground>
-        </div>
+            )}
+        </MobileContext.Consumer>
     );
 }
