+++ src/Sponsors/Sponsors.tsx
@@ -7,7 +7,7 @@ import PageFade from "../Utils/PageFade";
 import GoldDivider from "../Utils/GoldDivider";
 import { ContentBackground } from "../Utils/ContentBackground";
 import GreenLink from "../Utils/GreenLink";
-import { useCurrentWidth } from "react-socks";
+import { MobileContext } from "../Context/MobileContext";
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
@@ -22,80 +22,82 @@ const useStyles = makeStyles((theme: Theme) =>
 export default function Sponsors() {
     const classes = useStyles();
 
-    const mobile = useCurrentWidth() < 1000;
-
     DocumentTitle({ title: "Sponsors" });
 
     return (
-        <PageFade>
-            <div style={{ margin: "auto", width: "95%" }}>
+        <MobileContext.Consumer>
+            {mobile => (
+                <PageFade>
+                    <div style={{ margin: "auto", width: "95%" }}>
 
-                <ContentBackground elevation={24} className={classes.container} >
-                    <Typography align="center" variant="h2" >
-                        Thank You To Our Sponsors For Their Support!
+                        <ContentBackground elevation={24} className={classes.container} >
+                            <Typography align="center" variant="h2" >
+                                Thank You To Our Sponsors For Their Support!
                     </Typography>
 
-                    <GoldDivider />
+                            <GoldDivider />
 
-                    <div className="levelSection" >
-                        <LevelCard level="Gold" />
+                            <div className="levelSection" >
+                                <LevelCard level="Gold" />
 
-                        <div className="sponsorCards" style={{ backgroundColor: "#EAD286" }} >
-                            <SponsorCard
-                                img={"https://img.pngio.com/walmart-logo-free-transparent-png-logos-walmart-logo-png-960_899.jpg"}
-                                orgName={"Walmart"}
-                                orgLink={"https://walmart.com/"}
-                            />
-                            <SponsorCard
-                                img={"http://imatrixsol.com/img/solidworks.png"}
-                                orgName={"SolidWorks"}
-                                orgLink={"https://www.solidworks.com/"}
-                            />
-                            <SponsorCard
-                                img={"https://retohercules.com/images/office-people-clipart-9.jpg"}
-                                orgName={"Your logo here!"}
-                            />
-                        </div>
+                                <div className="sponsorCards" style={{ backgroundColor: "#EAD286" }} >
+                                    <SponsorCard
+                                        img={"https://img.pngio.com/walmart-logo-free-transparent-png-logos-walmart-logo-png-960_899.jpg"}
+                                        orgName={"Walmart"}
+                                        orgLink={"https://walmart.com/"}
+                                    />
+                                    <SponsorCard
+                                        img={"http://imatrixsol.com/img/solidworks.png"}
+                                        orgName={"SolidWorks"}
+                                        orgLink={"https://www.solidworks.com/"}
+                                    />
+                                    <SponsorCard
+                                        img={"https://retohercules.com/images/office-people-clipart-9.jpg"}
+                                        orgName={"Your logo here!"}
+                                    />
+                                </div>
+                            </div>
+
+                            <div className="levelSection">
+                                <LevelCard level="Silver" />
+
+                                <div className="sponsorCards" style={{ backgroundColor: "#C1C0C0" }} >
+                                    <SponsorCard
+                                        img={"https://enventyspartners.com/wp-content/uploads/2017/09/cropped-en-favicon.gif"}
+                                        orgName={"Envyntus Partners"}
+                                        orgLink={"https://enventyspartners.com/"}
+                                    />
+                                    <SponsorCard
+                                        img={"https://retohercules.com/images/office-people-clipart-9.jpg"}
+                                        orgName={"Your logo here!"}
+                                    />
+                                </div>
+                            </div>
+
+                            <div className="levelSection">
+                                <LevelCard level="Bronze" />
+
+                                <div className="sponsorCards" style={{ backgroundColor: "#BDA780" }} >
+                                    <SponsorCard
+                                        img={"https://retohercules.com/images/office-people-clipart-9.jpg"}
+                                        orgName={"Your logo here!"}
+                                    />
+                                </div>
+                            </div>
+
+
+                        </ContentBackground>
+
+                        <ContentBackground elevation={24} className={classes.container} style={{ margin: "2rem auto", width: mobile ? "90%" : "60%" }}>
+                            <Typography align="center" variant="h3" style={{ lineHeight: mobile ? "2.5rem" : "3.5rem" }}>
+                                Ineterested in sponsoring?<br />Contact Sam Luu at<br />
+                                <GreenLink link="mailto:car-robotics@uncc.edu" text="car-robotics@uncc.edu" />
+                            </Typography>
+                        </ContentBackground>
+
                     </div>
-
-                    <div className="levelSection">
-                        <LevelCard level="Silver" />
-
-                        <div className="sponsorCards" style={{ backgroundColor: "#C1C0C0" }} >
-                            <SponsorCard
-                                img={"https://enventyspartners.com/wp-content/uploads/2017/09/cropped-en-favicon.gif"}
-                                orgName={"Envyntus Partners"}
-                                orgLink={"https://enventyspartners.com/"}
-                            />
-                            <SponsorCard
-                                img={"https://retohercules.com/images/office-people-clipart-9.jpg"}
-                                orgName={"Your logo here!"}
-                            />
-                        </div>
-                    </div>
-
-                    <div className="levelSection">
-                        <LevelCard level="Bronze" />
-
-                        <div className="sponsorCards" style={{ backgroundColor: "#BDA780" }} >
-                            <SponsorCard
-                                img={"https://retohercules.com/images/office-people-clipart-9.jpg"}
-                                orgName={"Your logo here!"}
-                            />
-                        </div>
-                    </div>
-
-
-                </ContentBackground>
-
-                <ContentBackground elevation={24} className={classes.container} style={{ margin: "2rem auto", width: mobile ? "90%" : "60%" }}>
-                    <Typography align="center" variant="h3" style={{ lineHeight: mobile ? "2.5rem" : "3.5rem" }}>
-                        Ineterested in sponsoring?<br />Contact Sam Luu at<br />
-                        <GreenLink link="mailto:car-robotics@uncc.edu" text="car-robotics@uncc.edu" />
-                    </Typography>
-                </ContentBackground>
-
-            </div>
-        </PageFade>
+                </PageFade>
+            )}
+        </MobileContext.Consumer>
     );
 }
